new API refactoring
