+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/api/server/CloseQueryHandler.java
@@ -18,7 +18,7 @@ package io.confluent.ksql.api.server;
 import static io.confluent.ksql.rest.Errors.ERROR_CODE_BAD_REQUEST;
 import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
 
-import io.confluent.ksql.api.server.protocol.CloseQueryArgs;
+import io.confluent.ksql.rest.entity.CloseQueryArgs;
 import io.vertx.core.Handler;
 import io.vertx.ext.web.RoutingContext;
 import java.util.Objects;
