+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/api/server/JsonQueryStreamResponseWriter.java
@@ -16,9 +16,8 @@
 package io.confluent.ksql.api.server;
 
 import io.confluent.ksql.GenericRow;
-import io.confluent.ksql.api.server.protocol.ErrorResponse;
-import io.confluent.ksql.api.server.protocol.PojoCodec;
-import io.confluent.ksql.api.server.protocol.QueryResponseMetadata;
+import io.confluent.ksql.rest.entity.KsqlErrorMessage;
+import io.confluent.ksql.rest.entity.QueryResponseMetadata;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpServerResponse;
 import java.util.Objects;
@@ -52,7 +51,7 @@ public class JsonQueryStreamResponseWriter implements QueryStreamResponseWriter 
   @Override
   public QueryStreamResponseWriter writeMetadata(final QueryResponseMetadata metaData) {
     final Buffer buff = Buffer.buffer().appendByte((byte) '[');
-    buff.appendBuffer(metaData.toBuffer());
+    buff.appendBuffer(PojoCodec.serializeObject(metaData));
     response.write(buff);
     return this;
   }
@@ -64,8 +63,8 @@ public class JsonQueryStreamResponseWriter implements QueryStreamResponseWriter 
   }
 
   @Override
-  public QueryStreamResponseWriter writeError(final ErrorResponse error) {
-    writeBuffer(error.toBuffer());
+  public QueryStreamResponseWriter writeError(final KsqlErrorMessage error) {
+    writeBuffer(PojoCodec.serializeObject(error));
     return this;
   }
