+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/api/server/QueryStreamResponseWriter.java
@@ -16,8 +16,8 @@
 package io.confluent.ksql.api.server;
 
 import io.confluent.ksql.GenericRow;
-import io.confluent.ksql.api.server.protocol.ErrorResponse;
-import io.confluent.ksql.api.server.protocol.QueryResponseMetadata;
+import io.confluent.ksql.rest.entity.KsqlErrorMessage;
+import io.confluent.ksql.rest.entity.QueryResponseMetadata;
 
 /**
  * Represents something that knows how to write out a query response
@@ -28,7 +28,7 @@ public interface QueryStreamResponseWriter {
 
   QueryStreamResponseWriter writeRow(GenericRow row);
 
-  QueryStreamResponseWriter writeError(ErrorResponse error);
+  QueryStreamResponseWriter writeError(KsqlErrorMessage error);
 
   void end();
