+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/api/server/ServerVerticle.java
@@ -15,38 +15,32 @@
 
 package io.confluent.ksql.api.server;
 
-import static io.confluent.ksql.rest.Errors.ERROR_CODE_FORBIDDEN;
-import static io.confluent.ksql.rest.Errors.ERROR_CODE_SERVER_ERROR;
-import static io.confluent.ksql.rest.Errors.ERROR_CODE_UNAUTHORIZED;
-import static io.netty.handler.codec.http.HttpResponseStatus.FORBIDDEN;
-import static io.netty.handler.codec.http.HttpResponseStatus.UNAUTHORIZED;
+import static io.confluent.ksql.api.server.OldApiUtils.handleOldApiRequest;
+import static io.netty.handler.codec.http.HttpResponseStatus.TEMPORARY_REDIRECT;
 
-import io.confluent.ksql.api.auth.AuthenticationPlugin;
-import io.confluent.ksql.api.auth.AuthenticationPluginHandler;
-import io.confluent.ksql.api.auth.JaasAuthProvider;
-import io.confluent.ksql.api.auth.KsqlAuthorizationProviderHandler;
-import io.confluent.ksql.api.server.protocol.ErrorResponse;
+import io.confluent.ksql.api.auth.ApiSecurityContext;
+import io.confluent.ksql.api.auth.DefaultApiSecurityContext;
 import io.confluent.ksql.api.spi.Endpoints;
-import io.confluent.ksql.rest.server.KsqlRestConfig;
-import io.confluent.ksql.security.KsqlSecurityExtension;
+import io.confluent.ksql.rest.entity.ClusterTerminateRequest;
+import io.confluent.ksql.rest.entity.HeartbeatMessage;
+import io.confluent.ksql.rest.entity.KsqlRequest;
+import io.confluent.ksql.rest.entity.LagReportingMessage;
+import io.confluent.ksql.rest.entity.Versions;
 import io.vertx.core.AbstractVerticle;
-import io.vertx.core.Handler;
 import io.vertx.core.Promise;
-import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpHeaders;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.http.HttpServer;
 import io.vertx.core.http.HttpServerOptions;
-import io.vertx.core.http.HttpServerResponse;
+import io.vertx.core.http.HttpServerRequest;
+import io.vertx.core.http.ServerWebSocket;
 import io.vertx.core.json.JsonObject;
-import io.vertx.ext.auth.AuthProvider;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
-import io.vertx.ext.web.handler.AuthHandler;
-import io.vertx.ext.web.handler.BasicAuthHandler;
 import io.vertx.ext.web.handler.BodyHandler;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -56,8 +50,12 @@ import org.slf4j.LoggerFactory;
  */
 // CHECKSTYLE_RULES.OFF: ClassDataAbstractionCoupling
 public class ServerVerticle extends AbstractVerticle {
+
   // CHECKSTYLE_RULES.ON: ClassDataAbstractionCoupling
   private static final Logger log = LoggerFactory.getLogger(ServerVerticle.class);
+
+  private static final String JSON_CONTENT_TYPE = "application/json";
+  private static final String DELIMITED_CONTENT_TYPE = "application/vnd.ksqlapi.delimited.v1";
 
   private final Endpoints endpoints;
   private final HttpServerOptions httpServerOptions;
@@ -114,173 +112,218 @@ public class ServerVerticle extends AbstractVerticle {
     router.route(HttpMethod.GET, "/chc/ready").handler(ServerVerticle::chcHandler);
     router.route(HttpMethod.GET, "/chc/live").handler(ServerVerticle::chcHandler);
 
-    PortedEndpoints.setupFailureHandler(router);
-
-    router.route().failureHandler(ServerVerticle::failureHandler);
+    router.route().failureHandler(new FailureHandler());
 
     isInternalListener.ifPresent(isInternal ->
         router.route().handler(new InternalEndpointHandler(isInternal)));
 
-    setupAuthHandlers(router);
+    AuthHandlers.setupAuthHandlers(server, router);
 
     router.route().handler(new ServerStateHandler(server.getServerState()));
 
+    // The new query and insert streaming API
+    // --------------------------------------
+
     router.route(HttpMethod.POST, "/query-stream")
-        .produces("application/vnd.ksqlapi.delimited.v1")
-        .produces("application/json")
+        .produces(DELIMITED_CONTENT_TYPE)
+        .produces(JSON_CONTENT_TYPE)
         .handler(BodyHandler.create())
         .handler(new QueryStreamHandler(endpoints, connectionQueryManager, context, server));
     router.route(HttpMethod.POST, "/inserts-stream")
-        .produces("application/vnd.ksqlapi.delimited.v1")
-        .produces("application/json")
+        .produces(DELIMITED_CONTENT_TYPE)
+        .produces(JSON_CONTENT_TYPE)
         .handler(new InsertsStreamHandler(context, endpoints, server.getWorkerExecutor()));
     router.route(HttpMethod.POST, "/close-query")
         .handler(BodyHandler.create())
         .handler(new CloseQueryHandler(server));
 
-    PortedEndpoints.setupEndpoints(endpoints, server, router);
+    // The old API which we continue to support as-is
+    // ----------------------------------------------
+
+    router.route(HttpMethod.GET, "/")
+        .handler(this::handleInfoRedirect);
+    router.route(HttpMethod.POST, "/ksql")
+        .handler(BodyHandler.create())
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleKsqlRequest);
+    router.route(HttpMethod.POST, "/ksql/terminate")
+        .handler(BodyHandler.create())
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleTerminateRequest);
+    router.route(HttpMethod.POST, "/query")
+        .handler(BodyHandler.create())
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleQueryRequest);
+    router.route(HttpMethod.GET, "/info")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleInfoRequest);
+    router.route(HttpMethod.POST, "/heartbeat")
+        .handler(BodyHandler.create())
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleHeartbeatRequest);
+    router.route(HttpMethod.GET, "/clusterStatus")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleClusterStatusRequest);
+    router.route(HttpMethod.GET, "/status/:type/:entity/:action")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleStatusRequest);
+    router.route(HttpMethod.GET, "/status")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleAllStatusesRequest);
+    router.route(HttpMethod.POST, "/lag")
+        .handler(BodyHandler.create())
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleLagReportRequest);
+    router.route(HttpMethod.GET, "/healthcheck")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleHealthcheckRequest);
+    router.route(HttpMethod.GET, "/v1/metadata")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleServerMetadataRequest);
+    router.route(HttpMethod.GET, "/v1/metadata/id")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleServerMetadataClusterIdRequest);
+    router.route(HttpMethod.GET, "/ws/query")
+        .produces(Versions.KSQL_V1_JSON)
+        .produces(JSON_CONTENT_TYPE)
+        .handler(this::handleWebsocket);
 
     return router;
   }
 
-  private static void failureHandler(final RoutingContext routingContext) {
-
-    log.error(String.format("Failed to handle request %d %s", routingContext.statusCode(),
-        routingContext.request().path()),
-        routingContext.failure());
-
-    final int statusCode = routingContext.statusCode();
-    final Throwable failure = routingContext.failure();
-
-    if (failure != null) {
-      if (failure instanceof KsqlApiException) {
-        final KsqlApiException ksqlApiException = (KsqlApiException) failure;
-        handleError(
-            routingContext.response(),
-            statusCode,
-            ksqlApiException.getErrorCode(),
-            ksqlApiException.getMessage()
-        );
-      } else {
-        final int errorCode;
-        if (statusCode == UNAUTHORIZED.code()) {
-          errorCode = ERROR_CODE_UNAUTHORIZED;
-        } else if (statusCode == FORBIDDEN.code()) {
-          errorCode = ERROR_CODE_FORBIDDEN;
-        } else {
-          errorCode = ERROR_CODE_SERVER_ERROR;
-        }
-        handleError(
-            routingContext.response(),
-            statusCode,
-            errorCode,
-            failure.getMessage()
-        );
-      }
-    } else {
-      routingContext.response().setStatusCode(statusCode).end();
-    }
+  private void handleKsqlRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, KsqlRequest.class,
+        (ksqlRequest, apiSecurityContext) ->
+            endpoints
+                .executeKsqlRequest(ksqlRequest, server.getWorkerExecutor(),
+                    DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private static void handleError(final HttpServerResponse response, final int statusCode,
-      final int errorCode, final String errMsg) {
-    final ErrorResponse errorResponse = new ErrorResponse(errorCode, errMsg);
-    final Buffer buffer = errorResponse.toBuffer();
-    response.setStatusCode(statusCode).end(buffer);
+  private void handleTerminateRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, ClusterTerminateRequest.class,
+        (request, apiSecurityContext) ->
+            endpoints
+                .executeTerminate(request, server.getWorkerExecutor(),
+                    DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private static void unhandledExceptionHandler(final Throwable t) {
-    log.error("Unhandled exception", t);
+  private void handleQueryRequest(final RoutingContext routingContext) {
+
+    final CompletableFuture<Void> connectionClosedFuture = new CompletableFuture<>();
+    routingContext.request().connection().closeHandler(v -> connectionClosedFuture.complete(null));
+
+    handleOldApiRequest(server, routingContext, KsqlRequest.class,
+        (request, apiSecurityContext) ->
+            endpoints
+                .executeQueryRequest(request, server.getWorkerExecutor(), connectionClosedFuture,
+                    DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private void setupAuthHandlers(final Router router) {
-    final Optional<AuthHandler> jaasAuthHandler = getJaasAuthHandler(server);
-    final KsqlSecurityExtension securityExtension = server.getSecurityExtension();
-    final Optional<AuthenticationPlugin> authenticationPlugin = server.getAuthenticationPlugin();
-    final Optional<Handler<RoutingContext>> pluginHandler =
-        authenticationPlugin.map(plugin -> new AuthenticationPluginHandler(server, plugin));
-
-    if (jaasAuthHandler.isPresent() || authenticationPlugin.isPresent()) {
-      router.route().handler(ServerVerticle::pauseHandler);
-
-      router.route().handler(rc -> wrappedAuthHandler(rc, jaasAuthHandler, pluginHandler));
-
-      // For authorization use auth provider configured via security extension (if any)
-      securityExtension.getAuthorizationProvider()
-          .ifPresent(ksqlAuthorizationProvider -> router.route()
-              .handler(new KsqlAuthorizationProviderHandler(server.getWorkerExecutor(),
-                  ksqlAuthorizationProvider)));
-
-      router.route().handler(ServerVerticle::resumeHandler);
-    }
+  private void handleInfoRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, null,
+        (request, apiSecurityContext) ->
+            endpoints.executeInfo(DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private static void wrappedAuthHandler(final RoutingContext routingContext,
-      final Optional<AuthHandler> jaasAuthHandler,
-      final Optional<Handler<RoutingContext>> pluginHandler) {
-    if (jaasAuthHandler.isPresent()) {
-      // If we have a Jaas handler configured and we have Basic credentials then we should auth
-      // with that
-      final String authHeader = routingContext.request().getHeader("Authorization");
-      if (authHeader != null && authHeader.toLowerCase().startsWith("basic ")) {
-        jaasAuthHandler.get().handle(routingContext);
-        return;
-      }
-    }
-    // Fall through to authing with any authentication plugin
-    if (pluginHandler.isPresent()) {
-      pluginHandler.get().handle(routingContext);
-    } else {
-      // Fail the request as unauthorized - this will occur if no auth plugin but Jaas handler
-      // is configured, but auth header is not basic auth
-      routingContext
-          .fail(UNAUTHORIZED.code(),
-              new KsqlApiException("Unauthorized", ERROR_CODE_UNAUTHORIZED));
-    }
+  private void handleClusterStatusRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, null,
+        (request, apiSecurityContext) ->
+            endpoints.executeClusterStatus(DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private static Optional<AuthHandler> getJaasAuthHandler(final Server server) {
-    final String authMethod = server.getConfig()
-        .getString(KsqlRestConfig.AUTHENTICATION_METHOD_CONFIG);
-    switch (authMethod) {
-      case KsqlRestConfig.AUTHENTICATION_METHOD_BASIC:
-        return Optional.of(basicAuthHandler(server));
-      case KsqlRestConfig.AUTHENTICATION_METHOD_NONE:
-        return Optional.empty();
-      default:
-        throw new IllegalStateException(String.format(
-            "Unexpected value for %s: %s",
-            KsqlRestConfig.AUTHENTICATION_METHOD_CONFIG,
-            authMethod
-        ));
-    }
+  private void handleHeartbeatRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, HeartbeatMessage.class,
+        (request, apiSecurityContext) ->
+            endpoints.executeHeartbeat(request, DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private static AuthHandler basicAuthHandler(final Server server) {
-    final AuthProvider authProvider = new JaasAuthProvider(server, server.getConfig());
-    final String realm = server.getConfig().getString(KsqlRestConfig.AUTHENTICATION_REALM_CONFIG);
-    final AuthHandler basicAuthHandler = BasicAuthHandler.create(authProvider, realm);
-    // It doesn't matter what we set here as we actually do the authorisation at the
-    // authentication stage and cache the result, but we must add an authority or
-    // no authorisation will be done
-    basicAuthHandler.addAuthority("ksql");
-    return basicAuthHandler;
+  private void handleStatusRequest(final RoutingContext routingContext) {
+    final HttpServerRequest request = routingContext.request();
+    final String type = request.getParam("type");
+    final String entity = request.getParam("entity");
+    final String action = request.getParam("action");
+    handleOldApiRequest(server, routingContext, null,
+        (r, apiSecurityContext) ->
+            endpoints.executeStatus(type, entity, action,
+                DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private static void pauseHandler(final RoutingContext routingContext) {
-    // prevent auth handler from reading request body
-    routingContext.request().pause();
-    routingContext.next();
+  private void handleAllStatusesRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, null,
+        (r, apiSecurityContext) ->
+            endpoints.executeAllStatuses(DefaultApiSecurityContext.create(routingContext))
+    );
   }
 
-  private static void resumeHandler(final RoutingContext routingContext) {
-    // Un-pause body handling as async auth provider calls have completed by this point
-    routingContext.request().resume();
-    routingContext.next();
+  private void handleLagReportRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, LagReportingMessage.class,
+        (request, apiSecurityContext) ->
+            endpoints.executeLagReport(request, DefaultApiSecurityContext.create(routingContext))
+    );
+  }
+
+  private void handleHealthcheckRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, null,
+        (request, apiSecurityContext) ->
+            endpoints.executeCheckHealth(DefaultApiSecurityContext.create(routingContext))
+    );
+  }
+
+  private void handleServerMetadataRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, null,
+        (request, apiSecurityContext) ->
+            endpoints.executeServerMetadata(DefaultApiSecurityContext.create(routingContext))
+    );
+  }
+
+  private void handleServerMetadataClusterIdRequest(final RoutingContext routingContext) {
+    handleOldApiRequest(server, routingContext, null,
+        (request, apiSecurityContext) ->
+            endpoints
+                .executeServerMetadataClusterId(DefaultApiSecurityContext.create(routingContext))
+    );
+  }
+
+  private void handleInfoRedirect(final RoutingContext routingContext) {
+    // We redirect to the /info endpoint.
+    // (This preserves behaviour of the old API)
+    routingContext.response().putHeader("location", "/info")
+        .setStatusCode(TEMPORARY_REDIRECT.code()).end();
+  }
+
+  private void handleWebsocket(final RoutingContext routingContext) {
+    final ApiSecurityContext apiSecurityContext = DefaultApiSecurityContext.create(routingContext);
+    final ServerWebSocket serverWebSocket = routingContext.request().upgrade();
+    endpoints
+        .executeWebsocketStream(serverWebSocket, routingContext.request().params(),
+            server.getWorkerExecutor(), apiSecurityContext);
   }
 
   private static void chcHandler(final RoutingContext routingContext) {
     routingContext.response().putHeader(HttpHeaders.CONTENT_TYPE.toString(), "application/json")
         .end(new JsonObject().toBuffer());
+  }
+
+  private static void unhandledExceptionHandler(final Throwable t) {
+    log.error("Unhandled exception", t);
   }
 }
