+++ ksqldb-rest-app/src/test/java/io/confluent/ksql/api/BaseApiTest.java
@@ -217,13 +217,12 @@ public class BaseApiTest {
 
   protected static void validateError(final int errorCode, final String message,
       final JsonObject error) {
-    assertThat(error.size(), is(5));
+    assertThat(error.size(), is(4));
     validateErrorCommon(errorCode, message, error);
   }
 
   protected static void validateErrorCommon(final int errorCode, final String message,
       final JsonObject error) {
-    assertThat(error.getString("status"), is("error"));
     assertThat(error.getInteger("error_code"), is(errorCode));
     assertThat(error.getString("message"), is(message));
   }
